requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4
mysql-connector-python
requests
python-dotenv
flask
